Multiple views support
* Missing code for offscreen view has been added in nfCore
* nfEngineDemo has been reorganized, new features:
  - "N" key spawns a new window.
  - "V" key shows/hides secondary camera view (renderer as an offscreen
    view).
  - "C" key changes secondary camera position and orientation.
  - Key "0"-"9" changes a scene only for current window. This way we can
    render multiple scenes to multiple windows or render the same scene
    multiple times to test various scenarios.

Note: This is not optimized yet. All scenes are rendered on a single
thread. Shadow maps are not reused if the same scene is drawn in many
views.

Verification: Rebuild all, run nfEngineDemo. Play with "N", "V", "C" and
"0"-"9" keys.

+++ nfEngineDemo/Editor.cpp
@@ -10,6 +10,7 @@ using namespace NFE::Math;
 using namespace NFE::Scene;
 using namespace NFE::Resource;
 
+EntityManager* gEntityManager = nullptr;
 EntityID gSelectedEntity = -1;
 
 // client data passed to AntTweakBar callbacks
@@ -24,7 +25,7 @@ enum class EditionType : int
 
 void TW_CALL Demo_EntitySetCallback(const void* value, void* clientData)
 {
-    if (value == nullptr)
+    if (value == nullptr || gEntityManager == nullptr)
         return;
 
     EditionType editionType = static_cast<EditionType>(reinterpret_cast<size_t>(clientData));
@@ -125,7 +126,7 @@ void TW_CALL Demo_EntitySetCallback(const void* value, void* clientData)
 
 void TW_CALL Demo_EntityGetCallback(void* value, void* clientData)
 {
-    if (value == nullptr)
+    if (value == nullptr || gEntityManager == nullptr)
         return;
 
     EditionType editionType = static_cast<EditionType>(reinterpret_cast<size_t>(clientData));
@@ -210,7 +211,7 @@ void TW_CALL Demo_EntityGetCallback(void* value, void* clientData)
 
 void TW_CALL Demo_EntityDeleteCallback(void* clientData)
 {
-    if (gSelectedEntity != -1)
+    if (gSelectedEntity != -1 && gEntityManager != nullptr)
     {
         gEntityManager->RemoveEntity(gSelectedEntity);
         gSelectedEntity = -1;
